# build: detect `inline` keyword