As I discovered during work on #12897, libcurl does not support use of `inline` keyword because libcurl can be built with unusual combination of old compiler and cmake.
When built by old compiler and autotools (configure), `inline` macro is defined to a suitable replacement (or empty string) if need.
The missing part is cmake checks for `inline` support.
This PR adds support for cmake detection of `inline`.
As a fallback path, `inline` compiler-time detection is implemented in `curl_setup.h`, so even if build system does not check for inline support for any reason the header will substitute a suitable replacement.

With this PR merged, libcurl and curl code can freely use `inline` keyword.