@bagder 2024-04-12
The configure script already checks for `inline` since 375cdf89ada. If not working, it adds a `#define inline` in `lib/curl_config.h` (ie replacing it with nothing). Maybe it would make sense to make the cmake check and logic mimic this?