@Karlson2k 2024-04-12
This is exactly what is implemented in this PR.
Plus workarounds if neither cmake nor configure is used