@bagder 2024-04-12
> This is exactly what is implemented in this PR.

Except:

>  Plus workarounds if neither cmake nor configure is used

... which I think you should drop.

We already use `inline` in libcurl code (see `lib/easy_lock.h`) and it works the way configure does it. It think you should limit this approach to only be for cmake and leave other build systems to handle the case themselves.