@vszakats 2024-04-12
What is the benefit of the CMake-level check compared to the detection done in C?