@vszakats 2024-04-13
I propose to move the inline feature check to C, by using this snippet:
```c
#ifndef CURL_INLINE
#if defined(__GNUC__) && (__GNUC__ >= 3)
#define CURL_INLINE __inline__
#elif defined(_MSC_VER) && (_MSC_VER >= 1400)
#define CURL_INLINE __inline
#elif (defined(__STDC_VERSION__) && (__STDC_VERSION__ >= 199901L)) || \
  (defined(__cplusplus) && (__cplusplus >= 199711L))
#define CURL_INLINE inline
#endif
#endif
```

This covers the vast majority of compilers without the complexity (and build performance drop) of this PR.
It also makes the code behave the same for all build systems. And it also allows customizing this by the builder via `-DCURL_INLINE=....

According to <https://github.com/nemequ/hedley/blob/8fb0604a8095f6c907378cc3f0391520ae843f6f/hedley.h#L1528-L1553> (logic also re-used by Brotli) there exists a few further exotic compilers, but I suggest explicitly adding them when curl actually supports those, and if a specific need arises.

This needs to adapt the single place where `inline` is used in curl, like this (plus applying the macro rename to the force inline logic above):
```
--- curl_sha512_256.c-ORI	2024-04-13 11:06:36.488362803 +0200
+++ curl_sha512_256.c	2024-04-13 11:06:56.795894365 +0200
@@ -338,7 +338,7 @@
 /* Defined as a function. The macro version may duplicate the binary code
  * size as each argument is used twice, so if any calculation is used
  * as an argument, the calculation could be done twice. */
-static MHDX_INLINE curl_uint64_t
+static CURL_INLINE curl_uint64_t
 MHDx_rotr64(curl_uint64_t value, unsigned int bits)
 {
   bits %= 64;
```