@Karlson2k 2024-04-13
> > Plus workarounds if neither cmake nor configure is used
> 
> ... which I think you should drop.
> 
> We already use `inline` in libcurl code (see `lib/easy_lock.h`) and it works the way configure does it. It think you should limit this approach to only be for cmake and leave other build systems to handle the case themselves.

Right. The only `inline` is in `easy_lock.h`, which looks like never compiled with MSVC. MSVC is the only old compiler checked with cmake.
If CI would had runners with old compilers AND cmake, they would fail on `easy_lock.h`.
I think this is the reason why `inline` is not used in curl code: every attempt to use `inline` faced errors from cmake + old MSVC builds and the easiest solution was just to drop `inline` so nobody have used it in PRs and patches.

It looks like libcurl and curl still could be built without autotools and cmake. See `/plan9`, `/winbuild`, `/packages/vms`, `/packages/OS400`. So if any external build system would be used, the code always would have fallback that should prevent compiler from failing.

> What is the advantage of the CMake-level check compared to the detection done in C?

It's easy: C macro detection is limited and imprecise, to be on the safe side it prefers to drop `inline` when it is not possible to precisely understand what is supported as there is no fallback path. As the result, for some of the compilers supporting `inline` (in one form or another), the `inline` will not be used (if only C-macro checks are used).

CMake detection always gives the precise result. Moreover, when standard `inline` is not supported cmake detects supported alternatives and substitute them if needed. The cmake detection does not rely on any kind of heuristic guessing.
Just like `configure` counterpart, which provides precise result as well.

> With this patch, curl would have 4 `inline` detections: in autotools (built-in), CMake, `curl_setup.h` and [`curl_sha512_256.c`](https://github.com/curl/curl/blob/de66e8ad38ba2fccbe965fdea909c7123c99315d/lib/curl_sha512_256.c#L294-L312).

Just three detection: CMake and autotools (the main build systems), `curl_setup.h` (in case if build system is not used for any reason). `curl_sha512_256.c` detects supported forms of `force inline`. The checks in `curl_sha512_256.c` could be greatly simplified, it was not include to keep the PR size minimal. 
I can add it for clarity.