@vszakats 2024-04-13
Is there an evidence that the "more precise" detection in CMake/autotools has actual benefits that justifies the complexity and build-performance hit it causes?

I'd like learn which common case is missed by the proposed alternative in https://github.com/curl/curl/pull/13355#issuecomment-2053584694. Once we know those, they are fairly simple to add.

> If CI would had runners with old compilers AND cmake, they would fail on easy_lock.h.

Above proposal fixes that too by replacing that `inline` with `CURL_INLINE`
and using the short detection.

Unless we're talking about a genuinly old or exotic compiler hitting that
`#if` branch, `easy_lock.h` is built without issues when using CMake on
Linux. See the recently added `curl-for-win / linux-musl-llvm` CI job, or 
`dist / verify-out-of-tree-cmake` (I haven't double-checked the latter).