@Karlson2k 2024-04-13
@vszakats I curl would care about the most common compilers, then it would be no need for any checks for `inline` as all common compilers support it natively nowadays. 

If you want to simplify the solution, then just use
``` C
#undef inline
#define inline /* empty */
```
The would cover 100% of all cases, exotic or not.
Very simple.

The current proposed solution just balances autotools and cmake builds, which is desired, as I understand (at least there is a special CI runner that compare results for autotools and cmake, however it is pretty limited as the only one compiler and configuration checked).
This PR provides guaranteed result and uses all available compiler features. This cover any untested and not directly supported compilers even with non-standard settings (like enforced strict C89 with long long).
You do not need to care about any missed (and not checked) compiler. Everything is handled automatically in the best possible way.