@vszakats 2024-04-13
You didn't answer the question of what is the benefit of going overboard with detecting a widely supported compiler feature using 4 complex methods, of which 2 are slow, and 2 are almost exact copies of each other, also most likely unnecessary if the CMake and autotools methods are 100% perfect as you say.

This, over a simple and easy to customize method that covers likely 99.99% of systems out there.

Also noting that this compiler feature is used ONCE in the whole codebase (only for non-OpenSSL builds)

(The two pre-existing bare uses in `easy_lock.h` did not cause any problem or reports since they were introduced.)

To put it differently, why is curl needing this elaborate detection dance, when Brotli is happy with a simple preprocessor detection, while also using `inline` much more for it's similarly portable and much more CPU-bound code?