@Karlson2k 2024-04-18
> * slows down builds. The data here is that each feature detection in CMake takes 2-3 seconds in CI (autotools probably similar). Equivalent to adding a new curl source file to src or lib. Some of these can be optimized out for Windows with a considerable effort, but it's just easier to avoid doing it.
>   The implemented CMake method uses 4 of these, worst case. This is 12 seconds. With 35 lines and custom C files compiled in each, it is complex.

Your numbers and assumptions are obviously wrong.
I made several runs of CMake configuration with clean cache under Windows.
* With this patch:  `[CMake] -- Configuring done (14.5s)`
* Without this patch:  `[CMake] -- Configuring done (14.0s)`

Stable numbers, best 5 out of 10 tries (the other five are 0.2-0.6 seconds longer, randomly).
The added delay is just 0.5 second. The whole CMake reconfiguration takes just 14 seconds.
Obviously, on modern compilers the only first case (with `inline`) is tried.
If the compiler is so old that it does not support any form of `inline`, then probably the performance is not something that user care about.

I also tried `configure` under the GNU/Linux.
Added command to print exact timestamps before and after the test.
The results:
``` log
2024-04-18 21:26:10.816365470+03:00
checking for inline... inline
2024-04-18 21:26:10.846448242+03:00
```
``` log
2024-04-18 21:26:25.190701707+03:00
checking for inline... inline
2024-04-18 21:26:25.237320318+03:00
```
``` log
2024-04-18 21:27:28.061953377+03:00
checking for inline... inline
2024-04-18 21:27:28.090844360+03:00
```

**The "slow down" is 0.03 seconds.**

Do you really think this critical slow down needs to be optimised?

> Equivalent to adding a new curl source file to src or lib.

I think it's obvious that a two-lines file without any includes (the test file) is compiled much faster than a full additional C file for curl or lib.

> * there was no evidence that in this case a reasonable detection couldn't be done in C.
> * in fact there was evidence to the contrary, e.g. brotli.

Could you please explain, how the code used in some other project proves some particular needs or absence of the particular needs for this project?